Revert "Martian of Week Update"
