+++ contribs/drt/src/main/java/org/matsim/contrib/drt/analysis/DrtVehicleDistanceStats.java
@@ -99,7 +99,7 @@ public class DrtVehicleDistanceStats
 	}
 
 	private void initializeVehicles() {
-		int maxCapacity = DrtAnalysisControlerListener.findMaxVehicleCapacity(fleetSpecification);
+		int maxCapacity = DrtLegsAnalyser.findMaxVehicleCapacity(fleetSpecification);
 		fleetSpecification.getVehicleSpecifications()
 				.keySet()
 				.stream()
