+++ contribs/hybridsim/pom.xml
@@ -10,8 +10,8 @@
   <name>hybridsim</name>
 
   <properties>
-    <protobuf.version>3.22.0</protobuf.version>
-    <grpc.version>1.53.0</grpc.version>
+    <protobuf.version>3.21.12</protobuf.version>
+    <grpc.version>1.52.1</grpc.version>
   </properties>
 
   <dependencies>
