+++ contribs/noise/src/main/java/org/matsim/contrib/noise/ReflectionContext.java
@@ -184,12 +184,6 @@ public class ReflectionContext {
 
     private Coordinate getReflectionSegment(Coordinate coordinate, LineSegment segment, double length) {
 
-		if (segment.getLength() == 0) {
-			// yyyy added this here to avoid crashing feb'2023 - cr
-			logger.warn("Zero length line segment on {}", coordinate);
-			return null;
-		}
-
         final Coordinate right = segment.pointAlongOffset(0.5, length);
         LineSegment segmentRight = new LineSegment(coordinate, right);
