+++ contribs/protobuf/pom.xml
@@ -11,7 +11,7 @@
     <name>protobuf</name>
 
     <properties>
-        <protobuf.version>3.22.0</protobuf.version>
+        <protobuf.version>3.21.12</protobuf.version>
     </properties>
 
     <dependencies>
